fix: Staging symbol mismatch fixes for objdiff alignment
- Fix 285 staging return values (GetDebugText string returns, etc.)
- Fix bool false/true returns with correct asm encoding
- Fix per-TU header tricks for base/derived return type mismatches
- Fix encoding fixes and bit extraction implementations (+16 matches)
- Correct return values and register encodings in staging stubs (+58 matches)
- Add staging_class_fwd.h for struct/class forward declarations
- Generate scalar deleting destructor staging files (+228 matches)

+++ src/staging/BaseInfo.000.cpp
@@ -1,5 +1,5 @@
 #include "BaseInfo.h"
 
-void GBaseInfo::UpdateValue(float param_1, uint32_t param_2, uint32_t param_3)
+void GBaseInfo::UpdateValue(float param_1, unsigned long param_2, unsigned long param_3)
 {
 }
