+++ src/staging/runblack.reassemble.0442.005405b0-00542510.cpp
@@ -1,7 +1,7 @@
 #include "LoginBox.h"
 
 // win1.41 00541040 mac 105beb60 LoginBox::Init(unsigned long, unsigned long, void (*)(int, SetupBox *, SetupControl *, int, int))
-void LoginBox::Init(uint32_t param_1, uint32_t param_2, DialogBoxInitCallback param_3)
+void LoginBox::Init(unsigned long param_1, unsigned long param_2, DialogBoxInitCallback param_3)
 {
 }
