+++ src/staging/runblack.reassemble.0456.005445b0-00544f80.cpp
@@ -2,7 +2,7 @@
 #include "SkirmishGameBox.h"
 
 // win1.41 005445b0 mac 105bae70 StartGameBox::Init(unsigned long, unsigned long, void (*)(int, SetupBox *, SetupControl *, int, int))
-void StartGameBox::Init(uint32_t param_1, uint32_t param_2, DialogBoxInitCallback param_3)
+void StartGameBox::Init(unsigned long param_1, unsigned long param_2, DialogBoxInitCallback param_3)
 {
 }
 
@@ -17,7 +17,7 @@ void StartGameBox::InitControls()
 }
 
 // win1.41 00544820 mac 105baab0 SkirmishGameBox::Init(unsigned long, unsigned long, void (*)(int, SetupBox *, SetupControl *, int, int))
-void SkirmishGameBox::Init(uint32_t param_1, uint32_t param_2, DialogBoxInitCallback param_3)
+void SkirmishGameBox::Init(unsigned long param_1, unsigned long param_2, DialogBoxInitCallback param_3)
 {
 }
