+++ src/staging/runblack.reassemble.0467.00547ff0-005497f0.cpp
@@ -1,7 +1,7 @@
 #include "SetupOnlineLandscapes.h"
 
 // win1.41 00547ff0 mac 105cbb10 SetupOnlineLandscapes::Init(unsigned long, unsigned long, void (*)(int, SetupBox *, SetupControl *, int, int))
-void SetupOnlineLandscapes::Init(uint32_t param_1, uint32_t param_2, DialogBoxInitCallback param_3)
+void SetupOnlineLandscapes::Init(unsigned long param_1, unsigned long param_2, DialogBoxInitCallback param_3)
 {
 }
