+++ src/staging/runblack.reassemble.0679.0061b880-0061e640.cpp
@@ -1,7 +1,7 @@
 #include "ChannelBox.h"
 
 // win1.41 0061bcd0 mac 1038f640 ChannelBox::Init(unsigned long, unsigned long, void (*)(int, SetupBox *, SetupControl *, int, int))
-void ChannelBox::Init(uint32_t param_1, uint32_t param_2, DialogBoxInitCallback param_3)
+void ChannelBox::Init(unsigned long param_1, unsigned long param_2, DialogBoxInitCallback param_3)
 {
 }
