+++ src/staging/runblack.reassemble.0705.00629740-0062ac00.cpp
@@ -7,7 +7,7 @@ void MPFEChannelSelector::RefreshChannelList()
 }
 
 // win1.41 0062a630 mac 1039bd40 MPFECreateGame::Init(unsigned long, unsigned long, void (*)(int, SetupBox *, SetupControl *, int, int))
-void MPFECreateGame::Init(uint32_t param_1, uint32_t param_2, DialogBoxInitCallback param_3)
+void MPFECreateGame::Init(unsigned long param_1, unsigned long param_2, DialogBoxInitCallback param_3)
 {
 }
